COMMERCE-11517 Handle fetch calls

+++ modules/apps/headless/headless-admin-user/headless-admin-user-impl/src/main/java/com/liferay/headless/admin/user/internal/resource/v1_0/AccountGroupResourceImpl.java
@@ -88,8 +88,10 @@ public class AccountGroupResourceImpl extends BaseAccountGroupResourceImpl {
 				_accountResourceDTOConverter.getAccountEntryId(
 					accountExternalReferenceCode));
 
-		_accountGroupRelService.deleteAccountGroupRel(
-			accountGroupRel.getAccountGroupRelId());
+		if (accountGroupRel != null) {
+			_accountGroupRelService.deleteAccountGroupRel(
+				accountGroupRel.getAccountGroupRelId());
+		}
 	}
 
 	@Override
